Move EnglishDatetimeUtilityConfiguration class to utilities folder
